Dumped ClientWorker
And renamed MemberWorker to Worker.
