--- simulator/src/test/java/com/hazelcast/simulator/worker/ClientWorkerTest.java
@@ -1,13 +1,0 @@
-package com.hazelcast.simulator.worker;
-
-import org.junit.Test;
-
-import static com.hazelcast.simulator.utils.ReflectionUtils.invokePrivateConstructor;
-
-public class ClientWorkerTest {
-
-    @Test
-    public void testConstructor() throws Exception {
-        invokePrivateConstructor(ClientWorker.class);
-    }
-}
